test: isolate bootstrap init helpers from test files
